refactor(vm): split vm_react_loop.rs into cohesive submodules (§7-8)
`vm_react_loop.rs` (970 lines) exceeded the 500-line module convention. Split
into three cohesive files, byte-identically (function multiset verified
unchanged; methods reached via `self`, visibility preserved by bumping the
cross-file private helpers to `pub(super)`):

- `vm_react_loop.rs` (398): event-loop core — callback dispatch
  (`call_react_callback`), supply QUIT handler, the drain/consumer helpers, and
  the main `run_react_event_loop`.
- `vm_react_subscriptions.rs` (352): the shared `drive_react_subscriptions_nested`
  loop backing both `react {}` and the `await $supply` / `.Promise` paths.
- `vm_react_supply_helpers.rs` (246): replay + inner-supply drive helpers
  (`replay_static_supply`, `drive_inner_supply_to_consumer`, head-limit /
  on-close-callback / channel-id extraction).

All three files are now under 500 lines.
