+++ src/vm/vm_react_loop.rs
@@ -23,12 +23,10 @@
 
 use super::*;
 use crate::runtime::native_methods::{
-    SupplyEvent, next_supplier_id, supplier_register_promise, supplier_snapshot,
-    take_promise_combinator_sources, take_supply_channel,
+    SupplyEvent, next_supplier_id, supplier_register_promise, take_supply_channel,
 };
 use crate::runtime::subtest::{ReactSubscription, StreamConsumer, SupplyDrivePolicy};
 use std::sync::mpsc;
-use std::time::Duration;
 
 impl Interpreter {
     /// Dispatch a `whenever` body or one of its `LAST` / `QUIT` / `CLOSE` phaser
@@ -40,7 +38,11 @@ impl Interpreter {
     /// drop the topic. Loop-control signals (`done` / `next` / `last`) still
     /// surface as `Err` exactly as the tree-walk path produced them, so the
     /// drive loop's signal mapping (`run_react_consumer` etc.) is unchanged.
-    fn call_react_callback(&mut self, cb: &Value, args: Vec<Value>) -> Result<Value, RuntimeError> {
+    pub(super) fn call_react_callback(
+        &mut self,
+        cb: &Value,
+        args: Vec<Value>,
+    ) -> Result<Value, RuntimeError> {
         let topic = args.first().cloned();
         self.vm_call_map_block(cb, args, topic, false)
     }
@@ -96,7 +98,7 @@ impl Interpreter {
     ///   The react keeps driving any other subscriptions.
     ///
     /// Any other error propagates.
-    fn run_react_consumer(
+    pub(super) fn run_react_consumer(
         &mut self,
         sub: &mut ReactSubscription,
         value: Value,
@@ -393,578 +395,5 @@ impl Interpreter {
         }
 
         self.drive_react_subscriptions_nested(react_subs, SupplyDrivePolicy::React)
-    }
-
-    /// Shared subscription drive loop backing both `react { ... }` and the
-    /// `await $supply` / `$supply.Promise` paths. `react`-built and
-    /// promise-built subscriptions poll through here; `policy` selects how each
-    /// emitted value is dispatched and when the loop completes (see
-    /// [`SupplyDrivePolicy`]).
-    pub(crate) fn drive_react_subscriptions_nested(
-        &mut self,
-        mut react_subs: Vec<ReactSubscription>,
-        mut policy: SupplyDrivePolicy,
-    ) -> Result<(), RuntimeError> {
-        if react_subs.is_empty() {
-            if let SupplyDrivePolicy::Promise {
-                promise,
-                last_value,
-                ..
-            } = &policy
-                && !promise.is_resolved()
-            {
-                promise.keep(last_value.clone(), String::new(), String::new());
-            }
-            return Ok(());
-        }
-
-        // Event loop: poll all subscriptions
-        let timeout = Duration::from_millis(10);
-        'react_loop: loop {
-            if let SupplyDrivePolicy::Promise {
-                promise, deadline, ..
-            } = &policy
-            {
-                // The supply body's inner `done` keeps this promise through the
-                // supplier registry; once that happens the await is satisfied.
-                if promise.is_resolved() {
-                    return Ok(());
-                }
-                // Bound the wait so a stalled source cannot hang the await.
-                if std::time::Instant::now() >= *deadline {
-                    promise.keep(Value::Nil, String::new(), String::new());
-                    return Ok(());
-                }
-            }
-            let mut all_done = true;
-            for sub in react_subs.iter_mut() {
-                if sub.done {
-                    continue;
-                }
-                all_done = false;
-                // On-demand supply completion: a `done` inside the `supply { ... }`
-                // body was rewritten to `$emitter.done()`, which resolves this
-                // subscription's done-signal promise rather than raising the
-                // react-done signal. Once that resolves the flattened subscription
-                // is complete, so fire its LAST callbacks and stop polling it
-                // (otherwise an infinite source like `Supply.interval` would spin
-                // forever — see S17-supply/syntax.t).
-                if let Some(done_promise) = sub.on_demand_done.clone()
-                    && done_promise.is_resolved()
-                {
-                    for callback in &sub.last_callbacks {
-                        match self.call_react_callback(&callback.clone(), Vec::new()) {
-                            Err(e) if e.is_react_done() => break 'react_loop,
-                            other => {
-                                other?;
-                            }
-                        }
-                    }
-                    sub.done = true;
-                    continue;
-                }
-                // Handle Channel sources: poll values directly
-                if let Some(ref ch) = sub.channel {
-                    match ch.poll_result() {
-                        Ok(Some(value)) => {
-                            if self.run_react_consumer(sub, value)? {
-                                break 'react_loop;
-                            }
-                            sub.emit_count += 1;
-                        }
-                        Ok(None) => {
-                            // No value available yet; if channel is closed, mark done
-                            if !ch.can_send() {
-                                for callback in &sub.last_callbacks {
-                                    self.call_react_callback(&callback.clone(), Vec::new())?;
-                                }
-                                sub.done = true;
-                            }
-                        }
-                        Err(_err) => {
-                            sub.done = true;
-                        }
-                    }
-                    continue;
-                }
-                if let Some(supplier_id) = sub.supplier_id {
-                    let (values, done, quit) = supplier_snapshot(supplier_id);
-                    while sub.supplier_next_index < values.len() {
-                        // Check head_limit before processing more values
-                        if let Some(limit) = sub.head_limit
-                            && sub.emit_count >= limit
-                        {
-                            // Reached the head limit — mark as done
-                            for callback in &sub.last_callbacks {
-                                self.call_react_callback(&callback.clone(), Vec::new())?;
-                            }
-                            sub.done = true;
-                            break;
-                        }
-                        let value = values[sub.supplier_next_index].clone();
-                        sub.supplier_next_index += 1;
-                        if sub.is_lines {
-                            let chunk = value.to_string_value();
-                            sub.line_buffer.push_str(&chunk);
-                            while let Some(pos) = sub.line_buffer.find('\n') {
-                                let line = sub.line_buffer[..pos].to_string();
-                                sub.line_buffer = sub.line_buffer[pos + 1..].to_string();
-                                if self.run_react_consumer(sub, Value::str(line))? {
-                                    break 'react_loop;
-                                }
-                                if sub.done {
-                                    break;
-                                }
-                                sub.emit_count += 1;
-                                if let Some(limit) = sub.head_limit
-                                    && sub.emit_count >= limit
-                                {
-                                    for callback in &sub.last_callbacks {
-                                        self.call_react_callback(&callback.clone(), Vec::new())?;
-                                    }
-                                    sub.done = true;
-                                    break;
-                                }
-                            }
-                        } else {
-                            if self.run_react_consumer(sub, value)? {
-                                break 'react_loop;
-                            }
-                            // `last` in the body marked this whenever done; stop
-                            // pulling further values from the source.
-                            if sub.done {
-                                break;
-                            }
-                            sub.emit_count += 1;
-                            if let Some(limit) = sub.head_limit
-                                && sub.emit_count >= limit
-                            {
-                                for callback in &sub.last_callbacks {
-                                    self.call_react_callback(&callback.clone(), Vec::new())?;
-                                }
-                                sub.done = true;
-                                break;
-                            }
-                        }
-                    }
-                    if let Some(error) = quit {
-                        let mut handled = false;
-                        for quit_cb in &sub.quit_callbacks {
-                            self.call_supply_quit_handler(quit_cb.clone(), error.clone())?;
-                            handled = true;
-                        }
-                        if handled {
-                            sub.done = true;
-                            continue;
-                        }
-                        Self::run_react_close_callbacks(self, &react_subs);
-                        let quit_err =
-                            crate::runtime::Interpreter::runtime_error_from_supply_reason(error);
-                        return Err(crate::runtime::Interpreter::wrap_react_died(quit_err));
-                    }
-                    if done {
-                        if sub.is_lines && !sub.line_buffer.is_empty() {
-                            let remaining = std::mem::take(&mut sub.line_buffer);
-                            match self.call_react_callback(
-                                &sub.callback.clone(),
-                                vec![Value::str(remaining)],
-                            ) {
-                                Err(e) if e.is_react_done() => break 'react_loop,
-                                other => {
-                                    other?;
-                                }
-                            }
-                        }
-                        for callback in &sub.last_callbacks {
-                            self.call_react_callback(&callback.clone(), Vec::new())?;
-                        }
-                        sub.done = true;
-                    }
-                    continue;
-                }
-                let Some(receiver) = sub.receiver.as_ref() else {
-                    sub.done = true;
-                    continue;
-                };
-                if let Some(promise) = sub.promise.as_ref()
-                    && let Some(sources) = take_promise_combinator_sources(promise)
-                {
-                    for source in sources {
-                        source.result_blocking();
-                    }
-                    continue;
-                }
-                // Try to receive with a short timeout
-                match receiver.recv_timeout(timeout) {
-                    Ok(SupplyEvent::Emit(value)) => match &mut policy {
-                        SupplyDrivePolicy::Promise {
-                            promise,
-                            last_value,
-                            ..
-                        } => {
-                            // Capture values the whenever block `emit`s so a
-                            // later `done` resolves the promise with the last one.
-                            self.supply_emit_buffer.push(Vec::new());
-                            let cb_result =
-                                self.call_react_callback(&sub.callback.clone(), vec![value]);
-                            let emitted = self.supply_emit_buffer.pop().unwrap_or_default();
-                            if let Some(last) = emitted.last() {
-                                *last_value = last.clone();
-                            }
-                            if promise.is_resolved() {
-                                return Ok(());
-                            }
-                            if let Err(err) = cb_result {
-                                // `done`/`last` inside the whenever complete the
-                                // supply: keep the promise with the last emitted
-                                // value immediately rather than spinning to the
-                                // deadline.
-                                if err.is_react_done() || err.is_last() {
-                                    promise.keep(last_value.clone(), String::new(), String::new());
-                                    return Ok(());
-                                }
-                                // `next`/`redo` are loop control, not completion.
-                                if !err.is_next() && !err.is_redo() {
-                                    // A `die` quits the supply: break with the cause.
-                                    let cause = err
-                                        .exception
-                                        .as_deref()
-                                        .cloned()
-                                        .unwrap_or_else(|| Value::str(err.message.clone()));
-                                    promise.break_with(cause, String::new(), String::new());
-                                    return Ok(());
-                                }
-                            }
-                        }
-                        SupplyDrivePolicy::React => {
-                            if sub.is_lines {
-                                let chunk = value.to_string_value();
-                                sub.line_buffer.push_str(&chunk);
-                                while let Some(pos) = sub.line_buffer.find('\n') {
-                                    let line = sub.line_buffer[..pos].to_string();
-                                    sub.line_buffer = sub.line_buffer[pos + 1..].to_string();
-                                    if self.run_react_consumer(sub, Value::str(line))? {
-                                        break 'react_loop;
-                                    }
-                                    if sub.done {
-                                        break;
-                                    }
-                                }
-                            } else if self.run_react_consumer(sub, value)? {
-                                break 'react_loop;
-                            }
-                        }
-                    },
-                    Ok(SupplyEvent::Done) => {
-                        if matches!(policy, SupplyDrivePolicy::Promise { .. }) {
-                            // Inner supply done: the promise resolves through the
-                            // supplier registry, not the channel close — just
-                            // retire this receiver.
-                            sub.done = true;
-                        } else {
-                            if sub.is_lines && !sub.line_buffer.is_empty() {
-                                let remaining = std::mem::take(&mut sub.line_buffer);
-                                match self.call_react_callback(
-                                    &sub.callback.clone(),
-                                    vec![Value::str(remaining)],
-                                ) {
-                                    Err(e) if e.is_react_done() => break 'react_loop,
-                                    other => {
-                                        other?;
-                                    }
-                                }
-                            }
-                            for callback in &sub.last_callbacks {
-                                self.call_react_callback(&callback.clone(), Vec::new())?;
-                            }
-                            sub.done = true;
-                        }
-                    }
-                    Ok(SupplyEvent::Quit(error)) => {
-                        if matches!(policy, SupplyDrivePolicy::Promise { .. }) {
-                            // On the await path an inner quit just retires the
-                            // receiver; the promise is resolved/broken elsewhere.
-                            sub.done = true;
-                        } else {
-                            let mut handled = false;
-                            for quit_cb in &sub.quit_callbacks {
-                                self.call_supply_quit_handler(quit_cb.clone(), error.clone())?;
-                                handled = true;
-                            }
-                            sub.done = true;
-                            if !handled {
-                                let ch_quit_err =
-                                    crate::runtime::Interpreter::runtime_error_from_supply_reason(
-                                        error,
-                                    );
-                                return Err(crate::runtime::Interpreter::wrap_react_died(
-                                    ch_quit_err,
-                                ));
-                            }
-                        }
-                    }
-                    Err(mpsc::RecvTimeoutError::Timeout) => {}
-                    Err(mpsc::RecvTimeoutError::Disconnected) => {
-                        sub.done = true;
-                    }
-                }
-            }
-            match &policy {
-                SupplyDrivePolicy::React => {
-                    if all_done || react_subs.iter().all(|s| s.done) {
-                        break;
-                    }
-                }
-                SupplyDrivePolicy::Promise { promise, .. } => {
-                    if promise.is_resolved() {
-                        return Ok(());
-                    }
-                    // All channels closed without the promise resolving:
-                    // complete the await with Nil (matching the legacy
-                    // supply_promise_on_demand loop).
-                    if all_done || react_subs.iter().all(|s| s.done) {
-                        promise.keep(Value::Nil, String::new(), String::new());
-                        return Ok(());
-                    }
-                }
-            }
-        }
-
-        Self::run_react_close_callbacks(self, &react_subs);
-        Ok(())
-    }
-
-    fn extract_head_limit(attributes: &HashMap<String, Value>) -> Option<usize> {
-        if let Some(Value::Int(n)) = attributes.get("head_limit") {
-            Some(*n as usize)
-        } else {
-            None
-        }
-    }
-
-    fn extract_supply_on_close_callbacks(attributes: &HashMap<String, Value>) -> Vec<Value> {
-        if let Some(Value::Array(callbacks, ..)) = attributes.get("on_close_callbacks") {
-            callbacks.to_vec()
-        } else {
-            Vec::new()
-        }
-    }
-
-    fn run_react_close_callbacks(&mut self, react_subs: &[ReactSubscription]) {
-        for sub in react_subs {
-            for callback in &sub.close_callbacks {
-                let _ = self.call_react_callback(&callback.clone(), Vec::new());
-            }
-        }
-    }
-
-    /// Resolve the supply_id to use for channel lookup.
-    /// For a "lines" supply, use the parent_supply_id.
-    fn resolve_supply_channel_id(&self, attributes: &HashMap<String, Value>) -> Option<u64> {
-        // If this is a "lines" supply, use the parent's supply_id
-        if let Some(Value::Int(parent_id)) = attributes.get("parent_supply_id") {
-            return Some(*parent_id as u64);
-        }
-        // Otherwise use this supply's own supply_id
-        if let Some(Value::Int(id)) = attributes.get("supply_id") {
-            return Some(*id as u64);
-        }
-        None
-    }
-
-    /// Replay static supply values (non-streaming supplies)
-    /// Replay a static supply's values through a `whenever` callback, honouring
-    /// the loop-control signals the body may raise: `done` ends the react
-    /// (returns `Ok(true)`), `next` skips to the next value, `last` stops the
-    /// whenever early. This whenever's LAST phasers are fired here — with the
-    /// triggering value as topic when stopped via `last`, otherwise with no
-    /// topic on natural completion. Returns `Ok(true)` iff `done` was signalled.
-    /// Replay an inner `whenever` subscription registration
-    /// `[Supply, callback, last?, quit?]` whose source has no live channel or
-    /// supplier id (so `value_to_react_subscription` returned None) — a finite
-    /// `Supply.from-list` source, or a chained `supply { ... }` transform.
-    /// Returns `Some(true)` if a body signalled react `done`, `Some(false)` after
-    /// a normal replay, or `None` if `v` is not such a Supply registration.
-    fn replay_inner_static_subscription(
-        &mut self,
-        v: &Value,
-    ) -> Result<Option<bool>, RuntimeError> {
-        let Value::Array(inner_items, ..) = v else {
-            return Ok(None);
-        };
-        let Some(source @ Value::Instance { class_name, .. }) = inner_items.first() else {
-            return Ok(None);
-        };
-        if class_name != "Supply" {
-            return Ok(None);
-        }
-        let source = source.clone();
-        let inner_cb = inner_items.get(1).cloned().unwrap_or(Value::Nil);
-        let inner_last = inner_items
-            .get(2)
-            .and_then(crate::runtime::Interpreter::value_array_items)
-            .unwrap_or_default();
-        self.drive_inner_supply_to_consumer(&source, &inner_cb, &inner_last)
-            .map(Some)
-    }
-
-    /// Synchronously drive a finite `Supply` `source`, delivering each value to
-    /// `consumer_cb`. Handles two shapes that the polling event loop cannot:
-    /// a `values`-backed source (`Supply.from-list`) replayed directly, and a
-    /// chained `supply { whenever <upstream> { emit ... } }` transform whose
-    /// `whenever` source is itself an on-demand supply (recursed into). For the
-    /// on-demand case a streaming consumer is registered so the body's `emit`s
-    /// reach `consumer_cb` via `try_stream_emit`; inner registrations are driven
-    /// recursively, which is what lets a multi-stage Cro pipeline pass values all
-    /// the way through. Returns `true` if any body signalled react `done`.
-    fn drive_inner_supply_to_consumer(
-        &mut self,
-        source: &Value,
-        consumer_cb: &Value,
-        last_callbacks: &[Value],
-    ) -> Result<bool, RuntimeError> {
-        let Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } = source
-        else {
-            return Ok(false);
-        };
-        if class_name != "Supply" {
-            return Ok(false);
-        }
-        // Guard against runaway nesting. A pipeline that reuses the *same*
-        // `supply { ... }` transform sub more than once shares one emitter
-        // variable name (assigned per parse-site), which a deeper fix to lexical
-        // emitter capture would remove; until then a same-sub-twice chain can
-        // loop. Bail gracefully instead of overflowing the stack.
-        if self.supply_stream_consumers.len() > 256 {
-            return Err(RuntimeError::new(
-                "supply pipeline nested too deeply (a transform sub reused in the \
-                 same pipeline shares an emitter binding)",
-            ));
-        }
-        let attrs = attributes.as_map();
-        // A `supply { ... }` block carries both an (empty) `values` array and an
-        // `on_demand_callback`; a `Supply.from-list` carries only `values`. Prefer
-        // the on-demand body when present so a chained transform is driven, not
-        // mistaken for an empty static source.
-        let on_demand_cb = match attrs.get("on_demand_callback").cloned() {
-            Some(cb) => cb,
-            None => {
-                if attrs.contains_key("values") {
-                    return self.replay_static_supply(&attrs, consumer_cb, last_callbacks);
-                }
-                // A channel-backed or otherwise live source cannot be replayed here.
-                return Ok(false);
-            }
-        };
-        let sid = next_supplier_id();
-        self.supply_stream_consumers.push(StreamConsumer {
-            supplier_id: sid,
-            consumer_cb: consumer_cb.clone(),
-            done: false,
-        });
-        let idx = self.supply_stream_consumers.len() - 1;
-        let (res, emitted, _) = loan_env!(self, run_on_demand_body(on_demand_cb, Some(sid)));
-        let mut reached = self
-            .supply_stream_consumers
-            .get(idx)
-            .map(|c| c.done)
-            .unwrap_or(false);
-        if let Err(e) = res
-            && !e.is_react_done()
-        {
-            self.supply_stream_consumers.truncate(idx);
-            return Err(crate::runtime::Interpreter::wrap_react_died(e));
-        }
-        if !reached {
-            for v in emitted {
-                if crate::runtime::Interpreter::is_supply_subscription_registration(&v) {
-                    if let Value::Array(inner, ..) = &v
-                        && let Some(inner_src @ Value::Instance { .. }) = inner.first()
-                    {
-                        let inner_src = inner_src.clone();
-                        let inner_cb = inner.get(1).cloned().unwrap_or(Value::Nil);
-                        let inner_last = inner
-                            .get(2)
-                            .and_then(crate::runtime::Interpreter::value_array_items)
-                            .unwrap_or_default();
-                        if self.drive_inner_supply_to_consumer(
-                            &inner_src,
-                            &inner_cb,
-                            &inner_last,
-                        )? {
-                            reached = true;
-                            break;
-                        }
-                    }
-                } else {
-                    // A plain value buffered by the body (no active stream route):
-                    // deliver it to the consumer directly.
-                    match self.call_react_callback(&consumer_cb.clone(), vec![v]) {
-                        Err(e) if e.is_react_done() => {
-                            reached = true;
-                            break;
-                        }
-                        other => {
-                            other?;
-                        }
-                    }
-                }
-            }
-        }
-        self.supply_stream_consumers.truncate(idx);
-        if !reached {
-            for cb in last_callbacks {
-                match self.call_react_callback(&cb.clone(), Vec::new()) {
-                    Err(e) if e.is_react_done() => {
-                        reached = true;
-                        break;
-                    }
-                    other => {
-                        other?;
-                    }
-                }
-            }
-        }
-        Ok(reached)
-    }
-
-    fn replay_static_supply(
-        &mut self,
-        attributes: &HashMap<String, Value>,
-        callback: &Value,
-        last_callbacks: &[Value],
-    ) -> Result<bool, RuntimeError> {
-        let mut last_topic: Option<Value> = None;
-        if let Some(Value::Array(values, ..)) = attributes.get("values") {
-            for v in values.iter() {
-                match self.call_react_callback(&callback.clone(), vec![v.clone()]) {
-                    Ok(_) => {}
-                    Err(e) if e.is_react_done() => return Ok(true),
-                    Err(e) if e.is_next() => continue,
-                    Err(e) if e.is_last() => {
-                        last_topic = Some(v.clone());
-                        break;
-                    }
-                    Err(e) => return Err(e),
-                }
-            }
-        }
-        for cb in last_callbacks {
-            let args = match &last_topic {
-                Some(v) => vec![v.clone()],
-                None => Vec::new(),
-            };
-            match self.call_react_callback(&cb.clone(), args) {
-                Err(e) if e.is_react_done() => return Ok(true),
-                other => {
-                    other?;
-                }
-            }
-        }
-        Ok(false)
     }
 }
